blockstorage: Merge schedulerhints extension
This ones a little complicated since it's both an extension of the
Volume request struct and a validator for same.
