--- internal/acceptance/openstack/blockstorage/extensions/extensions.go
@@ -1,4 +1,0 @@
-// Package extensions contains common functions for creating block storage
-// resources that are extensions of the block storage API. See the `*_test.go`
-// files for example usages.
-package extensions
