--- internal/acceptance/openstack/blockstorage/extensions/pkg.go
@@ -1,3 +1,0 @@
-// The extensions package contains acceptance tests for the Openstack Cinder extensions service.
-
-package extensions
