+++ internal/acceptance/openstack/blockstorage/v3/schedulerhints_test.go
@@ -1,7 +1,7 @@
 //go:build acceptance || blockstorage
 // +build acceptance blockstorage
 
-package extensions
+package v3
 
 import (
 	"context"
@@ -10,7 +10,6 @@ import (
 
 	"github.com/gophercloud/gophercloud/v2/internal/acceptance/clients"
 	"github.com/gophercloud/gophercloud/v2/internal/acceptance/tools"
-	"github.com/gophercloud/gophercloud/v2/openstack/blockstorage/extensions/schedulerhints"
 	"github.com/gophercloud/gophercloud/v2/openstack/blockstorage/v3/volumes"
 	th "github.com/gophercloud/gophercloud/v2/testhelper"
 )
@@ -38,23 +37,18 @@ func TestSchedulerHints(t *testing.T) {
 	defer volumes.Delete(context.TODO(), client, volume1.ID, volumes.DeleteOpts{})
 
 	volumeName = tools.RandomString("ACPTTEST", 16)
-	base := volumes.CreateOpts{
-		Size: 1,
-		Name: volumeName,
-	}
-
-	schedulerHints := schedulerhints.SchedulerHints{
+	schedulerHints := volumes.SchedulerHints{
 		SameHost: []string{
 			volume1.ID,
 		},
 	}
-
-	createOptsWithHints := schedulerhints.CreateOptsExt{
-		VolumeCreateOptsBuilder: base,
-		SchedulerHints:          schedulerHints,
+	createOpts = volumes.CreateOpts{
+		Size:           1,
+		Name:           volumeName,
+		SchedulerHints: schedulerHints,
 	}
 
-	volume2, err := volumes.Create(context.TODO(), client, createOptsWithHints).Extract()
+	volume2, err := volumes.Create(context.TODO(), client, createOpts).Extract()
 	th.AssertNoErr(t, err)
 
 	ctx2, cancel2 := context.WithTimeout(context.TODO(), 60*time.Second)
