--- openstack/blockstorage/extensions/schedulerhints/testing/doc.go
@@ -1,2 +1,0 @@
-// schedulerhints unit tests
-package testing
